pyproject.toml
[project]
name = "boeing-737-rag"
version = "0.1.0"
description = "Advanced RAG API for Boeing 737 Technical Manual"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "unstructured[pdf]>=0.12.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.3.1",
    "rank-bm25>=0.2.2",
    "google-generativeai>=0.3.2",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "pypdf>=4.0.0",
    "tenacity",
    "FlagEmbedding>=1.2.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "ipykernel"
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

